Rename ErrorPage route module to PascalCase.
Align `src/` root components with naming conventions tracked in #1492.
